chore: implemented hot reloading of the echo server and the swagger docs
I changed the reflex.conf to restart the server when a go file was changed and to regenerate the swagger docs when a go file was changed. This is done by running the command `swag init && go run main.go` when a go file is changed.
I also changed the README.md to reflect the changes made to the reflex.conf file.