fix: avoid adding a minute to the timestamp
For some reason, when accepting a new inspector request without a timestamp, the timestamp is set to the current time plus a minute.
This caused the risk model to be out of sync with the recent inspectors as it always had to wait a minute before the new inspector was considered.