perf: call risk data less frequently
We were calling /getSegmentColors every 5 seconds to get the newest risk data.
This is now unnecessary, as we call it every time that we are getting a new inspector. But we still need to call it periodically to account for the risk that changes based on time.
But this is done over the span of an hour, so calling it every 5 seconds is a bit overkill.