refactor: station is unqiue to one line if unique to one of a type
If a station has mutliple lines but only one line of the same type of the given line,
then the station will marked as unique as the ticket inspectors cannot switch to another line of a different type.
So for instance if a given line was Ubahn and the station has 3 lines of Sbahn and 1 line of Ubahn, then we can mark the station as unique to one line.