refactor: extract displaying logic into a hook to avoid early termination due to isLoading getting set to false.
Before hand the issue was that the placeholder should get displayed for at least 1 second even if the loading has finished to avoid flickering. This did not work as the logic was internal in the component and once isLoading has been set to true it would unmount meaning the 1 second logic would not apply anymore as the order was comming from the parent and thus being more important. By using a hook we have seperated the logic from the rendering making it easier to controll when to render.