refactor: use handle_nlp_bot_error for all of the error handeling.
before we had three different ways of handeling the error, this redundancy makes it harder to debugg the code.
We are now using the handle_nlp_bot_error function for sending messages, logging etc. One issue remaining is that we still have an error watcher in the nlp bot and in the watcher bot that watches its sub process. So still some redundancy left