refactor: create a single story for multiple inspectors
the createImage and the createAndPostStory function now accept an array of inspectors instead of a single one, making it possible to make a single story for all of the current inspectors. This needs to be done as we can only make 25 posts a day, meaning we have to group them together.