refactor: only show skeleton again when station.id changes
The skeleton will be shown on initial load and whenever the station.id changes. When only userLat or userLng change, the old stationDistance will be displayed while the new distance is being fetched. The stationDistance will always be updated in the background when any of userLat, userLng, or station.id change.