feat: if no station, guess the station by using the line
There is an issue with a lot of reports not having a station id, so to tackle this I decided to guess the station id using the line and setting the most common station in the line as the station for the request. This is better than the status quo as alternatively we would return historic data which is even more in acurate.