feat: sort lines
the lines are now being sorted in a sensible way. The Ubahns are before the Sbahns and the S41, S42 are the first Sbahns. This will still make it easy to use the list as they are still sorted by their number but also some exceptions based on how often they are being reported. This should make it easier to report an inspector.