refactor: sort lines by their total risk
I modified `extractMostRiskLines()` to return the extracted lines from the `segment_colors` along with a type called `LineRiskData` which holds the score of a line and the class of a line.
The score is a value which is the accumulated `colorScores`, how high the total risk is and the class which is the hightest indivual score of a segment.

The Lines are being sorted based on this score, inorder to make the risk summary easier to read as the lines with the highest risk should be read first.