fix: also show not reported lines
We were creating the `riskMap` based on the lines that are mentioned in the `segment_colors` and the lines that were reported today.

This approach does not work as on some days some of the lines are not being checked.

I fixed this by looping over all of the lines not just todays reported lines.