feat: current layer in localstorage
Some users started to complain that they preferred the risk layer and that it was annoying to switch the layer every time.

I have solved this issue by changing the localstorage every time that we modify the current layer. While we also keep the state to make the app reactive without constantly checking the localstorage for a change.