fix: avoid showing most recent line report in top 5 (#216)
We were sorting all of the lines by their timestamp of the most recent
report and then displaying the first 5 as the top 5. This caused issues
as it would show the latest reports and not the top 5 reports but sorted
by most recent.

To fix this we are now sorting the lines by the number of their reports,
selecting the top 5 and then sorting them by their most recent report
again.