feat: compass that will rotate with the map (#219)
## Describe the issue:
If a user rotated the map there was no way to know where north was, this
made it difficult to navigate the map.

## Explain how you solved the issue:
Based on the suggestion by @rosahaj I have introduced a compass which
will rotate based on the maps rotation.
<img width="707" alt="Bildschirmfoto 2024-10-27 um 21 19 06"
src="https://github.com/user-attachments/assets/99b0eeac-ccad-40a1-8ac7-3ee92c1a5939">



## Additional notes or considerations:
I have added a custom 100% css icon because I found no icon that would
look good, but this makes the css for the compass a bit hacky. But I
think this is fine as the compass will not be changed a lot in the
future.