refactor: track inital layer and improved switching
We were only checking if the the risk layer was being changed to. This caused some issues when we have the risk layer as the default becasue we did not know how many people have selected it before.

We now have this info along with the info if they are switching back to the line layer.