refactor: single source of truth for line colors (#282)
This will make the UI less error prone as we now have a single source of
truth for the colors of the line.
This change was introduced because I forgot to update the colors of the
trams everywhere. It will now be harder for someone to forget to do
this.

## Notes and Considerations:
In a perfect wold we would also use the `getLineColor` in the
`RegularLineLayer` but I could not figure out how to get this to work.