refactor: minize resorting logic
Instead of tracking each layer change and manually moving the stationLayer an top we are now just moving the line layers below the station layer.

This will make the logic a lot more robust and fix the issue of the segments taking longer to load and thus being painted on top of the station layer without a switch in layer state.