feat: API serves segments
This is done to make it easier to change the segments in the future without needing to update mutliple json documents. This is especially crucial as updating them would require a review by the Appstore.

The results are being cached in a way in which we avoid fetching the segments as much as possible because the request is quite large.