Refactor theming with Restyle + design experimentation (#287)
## Describe the issue:
Previously, the app was using `native-base` for styling/themes. Besides
the fact that it is now deprecated, it had some dependency issues and
akward quirks that made using it for a proper design system a bit of a
pain.
## Explain how you solved the issue:
Switched to `@shopify/restyle`, which provides much more ergonomic
systems, and implemented unified themeing in the app
## Additional notes or considerations:
PR currently also contains a new colorscheme, however this is mostly for
experimental purposes and does not have to be released. Changing it
should be very easy with restyle.