refactor: show latest reports
instead of getting the most reported lines and then showing the latest reports of those. We are now just showing the latest reports that are grouped by lines. So even if a line has just a single report it can still be at the top of the list.

This was done as the former setup was confusing as only the top 5 lines were picked but they were then being sorted based on their latest report. So it was kind of doing two things at once.