feat: functional report button (#310)
## Describe the issue:
1 )While the feedback for the reports modal was really good the usage
was not as high as it could have been.
2) Some users had the app open for over 40 minutes during their entire
ride, checking if there was a new report. This was cumbersome as it was
hard to notice a new report if there are a lot of them on the map.
3) Some people wanted to see the total number of reports that are
currently active

## Explain how you solved the issue:
I changed the reports modal button to look like this:
<img width="372" alt="Bildschirmfoto 2024-11-27 um 12 27 26"
src="https://github.com/user-attachments/assets/c5b06297-b493-4c62-b2ba-519bfb50f527">

It does the following things:
- show the latest report
- highlight the latest report if it has not been viewed
- show the number of the reports on the map

This will allow people to easily see what is new with a simple glance.

## Additional notes or considerations:
I refactored the logic of the pulsing effect to also be used with the
opacity markers so that if you click them they wont be pulse anymore.