feat: segments with same start and endpoint inherit risk
this is being done to avoid the strange visual effect in which if two lines share the same segment. Like U1 and U3 but only one of them will have a report, then the risk for the other would be barely visible because it is being blocked by the non risk line.

Now if the U1 is reported the segments that the U1 and U3 share, the U3 will inherit the risk to avoid this visual effect.