refactor: rewrite risk model to python (#322)
## Describe the issue:
The risk model was written in r, this was causing massive issues for two
reasons:
a) r is a scripting language that is not designed to be used in a server
setting like this. Because of this we had a lot of overhead to manage a
simple script
b) no one of the core maintainers are familiar with r, meaning we had an
important piece of the app that was not being maintained. Due to this
changes were very hard and very risky.

## Explain how you solved the issue:
I rewrote the script in python, which is a lot easier to maintain and to
run. Instead of having a separate microservice we are now running the
python script directly on the server. We are passing the inputs and
output using stdin/stdout to avoid complexity and overhead.
With the current setup we can easily keep it as is or we can spin it out
to be its own microservice if the risk model grows.

Due to an issue with overlapping segments I also had to regenerate the
segments with a new overpass prompt. This made the segments file even
smaller :)
## Additional notes or considerations: