fix: keep old API contract
the old API still used if-modified-since caching. The new one uses server side in-memory caching and does not use any client side caching. Due to this I removed the property from the Response.

This caused the mobile app to break as it was still expecting it. I have now added it again with a placeholder value.

This will be removed again as soon as the app will update and finish the review.