feat: share marker with separate button (#332)
## Describe the issue:
In order to make the app more viral we need to constantly remind people
to share the app with their friends. This will improve the network
effects that this app is based on. Right now only a small number of
people clicked on the small share link inside of the marker modal.

## Explain how you solved the issue:
I added a bigger button that will prompt the user to share the app. The
disclaimer will now always show the message: "Daten vielleicht ungenau".
I also modified the visual hierarchy so that it will be easier to
comprehend the important info:
<img width="399" alt="Bildschirmfoto 2024-12-05 um 15 07 17"
src="https://github.com/user-attachments/assets/cfb2a3ea-1d02-430e-afcf-c47fb11f0799">

## Additional notes or considerations:
- I decided not have dynamic links for each inspector as this will add
minimal value becasue the info is already inside of the shared message
and after 60 minutes the marker wont be visible anymore so a lot of
people wont be able to see the report anyway once they click on the
link. And having dynamic routing would add complexity to the app which
we should avoid.