fix: overlapping segments inherit risk color
When two lines shared the same start and endpoint for a given segment, it would be very hard to see the risk color because one line was overlapping and thus blocking the color of the other line.

I have fixed this issue by propagating the risk to the overlapping segments. That way if there is a report on the S25 the S26 will also be colored for the shared segments.