chore: upgrade to react 19 (#341)
## Describe the issue:
A stable version of react 19 has been released. React 19 is recommended
if you want to make use of the react compiler.

## Additional notes or considerations:
After this is merged I will integrate the react-compiler eslint plugin
which will give us feedback like the compiler would do and once the
compiler is stable this can be fully integrated.