chore: setup linting config (#344)
## Describe the issue:
Our linting config was very liberal. This caused some issues as we made
some short cuts, that caused bugs which could have been caught by a
proper linting config.

@mclrc was kind enough to implement a linting config in #317. The issue
with this implementation was that a large number of changes were
necessary and some of these changes caused other things to break. This
would have made merging the PR very difficult and would block other
things.

## Explain how you solved the issue: 
I used the config that was set up by @mclrc and modified the CI workflow
to only check files that were changed in the PR. This will make it easy
to integrate the new changes becasue we can update the files one by one.

## Additional notes or considerations:
The linting config is a bit different from the mobile app. This is fine
as a bit of drit is okay for linting due to different requirements.