chore: configure react-compiler as eslint plugin (#347)
I have configured the compiler as an eslint plugin, this can be replaced
once the compiler it stable then it can be integrated into the build
pipeline.

I have also added the necessary react eslint plugins that will ensure we
are following the ["rules of react"](https://react.dev/reference/rules).
This will improve the compiler.