feat: track funnels (#349)
## Describe the issue:
It was unclear how many people clicked on links that were shared, for
example in the telegram group. This made it difficult to make decisions
around how to optimize them.

## Explain how you solved the issue:
I have added a special route that will send an analytics event that a
user has been funneled along with their source after this report is sent
the user will be redirected to app.freifahren.org. When the telegram bot
sends a message it will always send it with a link that is trackable.

This allows us to track the funnels without tracking any user specific
data such as their ip adress or so.

## Additional notes or considerations: