feat: avoid returning spam into telegram (#354)
## Describe the issue:
A user was submitting a large number of reports in a short time
(approximately 9 reports within one minute). This behavior is clearly
spam and poses a significant problem as these reports are transmitted
directly to the Telegram channel, making them visible to everyone. This
undermines the trust of Telegram admins, on whom we depend, as well as
the credibility of the system among users.

## Explain how you solved the issue:
To address this, I implemented a rate limit for Telegram reports. Now, a
message will only be sent if the previous one was more than 5 minutes
ago. This allows 95% of legitimate reports to pass through while
effectively blocking 100% of spam attempts.

## Additional notes or considerations:
While this solution is not perfect, it addresses the most urgent issue:
preserving trust in the system and with Telegram admins. Further
refinements may be needed in the future to improve the overall user
experience without compromising security.