fix: avoid race conditions with funnel event (#357)
## Describe the issue:
We were trying to track how many users clicked on the link from the
telegram group. This worked very seldomly.

## Explain how you solved the issue:
I avoided the race condition by waiting for pirsch sdk to be loaded and
only redirecting once this is done and the promise is returned.

Not 100% sure that this will work, lets see