chore: remove impressum
The german law requires all "Geschäftshähige" entities to have an impressum. Since we have no intention or option to make money, we are not "Geschäfshähig". This means we do not need an impressum.

It is better to remove the impressum if we dont need it, as we are exposing private information in the impressum and we are at risk of making false claims inside of the impressum that could get us into trouble.