chore: remove impressum (#367)
According to German law, only "geschäftsfähige" entities (those capable
of conducting business) are required to have an Impressum. Since our
project does not generate revenue and has no intention or capacity to do
so, we do not meet the criteria for being "geschäftsfähig." As a result,
we are not legally obligated to include an Impressum.

By removing the Impressum, we eliminate unnecessary exposure of private
information and reduce the risk of making unintentional false claims,
which could lead to potential legal issues. This change ensures that we
stay compliant while safeguarding personal information.