hotfix: Revert "feat: introduce server-side rate limiting" (#368)
Reverts FreiFahren/FreiFahren#365

I forgot that the rate limiting will also apply to the telegram bot.
This means that most of the reports that were coming from the bot were
being blocked. We will have to solve this in a seperate issue how we
introduce rate limiting without effecting the bot.