fix?: send funnel event in telegram browser
For some reason the telegram browser would not send the funnel clicked event but it would send all of the other events. This led me to suspect that it is because `App` was not yet mounted.

I have moved the event to be sent from `App` I hope that this will reliably send the event even from the telegram browser.