refactor: old `FeedbackForm` is `ContactSection`
This was done so that that the `ContactSection` can be embedded into the new `FeedbackForm` but the contact "page" can exist without the `FeedbackForm`.

I also used my full name. This is done so that the projects seems more legit if there is at least a single person who is identifiable.