feat: button to give feedback (#351)
## Describe the issue:
We had little qualitative feedback by users because you had to contact
us on telegram and the feedback button was not as prominent. This was a
massiv issue as we also have little quantiative feedback due to
regulations around privacy. Meaning we had little way of knowing if
changes are positive.

## Explain how you solved the issue:
I added a text input to the feedback modal so that users can give
feedback very quickly. I also added a feedback button in multiple
places.
<img width="200" alt="Bildschirmfoto 2024-12-31 um 14 18 29"
src="https://github.com/user-attachments/assets/a71403cd-ef52-48d0-b2eb-c5c157cfb737"
/> <img width="200" alt="Bildschirmfoto 2024-12-31 um 14 18 58"
src="https://github.com/user-attachments/assets/b3baebea-eca3-4370-8af7-dc02477e3238"
/>

When a user gives us feedback we have to get them to acknowledge the
privacy policy becasue they could be exposing personal data in the form.
The data is being sent to the backend where it is being saved in a table
of our database. I thought about sending the data to a google sheet, but
I decided against it because of legal issues regarding sending
potentially personal data to a non eu hosted google service.

When the feedback is submitted there will be a modal that will ask the
user to join the feedback telegram group. That way we will have a closer
contact with users and can get feedback quicker.
<img width="200" alt="Bildschirmfoto 2024-12-31 um 14 23 58"
src="https://github.com/user-attachments/assets/bd3ea49d-4511-4601-91a0-8afac8d67251"
/>

## Additional notes or considerations:
- I used my full name in the contact form so that the project will cause
more trust.
- I created a hook that will check if a form is valid. I did not get
around to also add this for the reportform but will do when I am working
on it the next time as this pr is already really big
- I also created the `useApi` abstraction layer so that we have a
central way of handling the errors. This will be very handy once we are
installing sentry.