feat: feat allow users to opt out of tracking (#372)
## Describe the Issue:
We were tracking anonymized user behavior. While this is legal and does
not pose any privacy concerns for users, it would be considerate to
provide users with the option to opt out of tracking. This demonstrates
our commitment to privacy and ensures users feel secure when using the
app.

## Explain How You Solved the Issue:
I added a button that stores the opt-out preference in `localStorage`.
Each time an analytics event is triggered, the app first checks whether
the user has opted out before sending the event.

## Additional Notes or Considerations:
I removed the light mode button since light mode will be removed
entirely very soon. Additionally, it was causing layout issues.