refactor: reportsmodal to be faster (#377)
## Describe the issue:
- When users opened the reports modal. It would show the reports of the
last hour in a clustered form, in which you could expand them to see the
reports of the last day as well. The previous approach fetched the data
from the last 24 hours when the user opend the modal. This caused it too
be slow as the user would have to wait for the request to finish
- @brandesdavid complained that the component was not easy to read.

## Explain how you solved the issue:
I moved the construction of the reportsList to the reportsContext. The
way it now works is that it will return the data from the last hour that
is already stored due to showing the markers it will then fetch the last
23 hours, concatinate the list with the data from the last hour and
return it again, refreshing the component.
That way we have a very accurate placeholder.

I also refactored a lot and split up the component.

## Additional Notes:
I removed the linting rule that enforces the catch errors when using a
promise. This is a programming anypattern imo as the cleaner solution is
to define errors out of existence inside of the abstraction layer. Else
we would have a lot of duplicate try catches.