Explanation of commit clean-and-fix-bot squash merge
## Describe the issue:

The bot had unhandled and implicit errors sometimes. The nlp bot thread, which was running in a separate thread, was not properly handled and would get stuck up. The logger did not log the errors, maybe due to the telebot library.
The restart functionality was not working as expected and in the end, not necessary.
The whole telegram bots folder was made as an abstraction layer. We don't need that anymore, but needed better error handling.

## Explain how you solved the issue:

added sentry
removed unnecessary restart utils
added connection to restart through api request

I added Sentry-Flask to the project to log them to sentry (See .env variable SENTRY_DSN)
I renamed the telegram bots folder to nlp_service.
I renamed every environment variables accordingly.
I removed the Restart_utils.py file. The restart functionality is now handled through the api request.
The coolify server is scheduled to curl inside the container to restart the bot, every 3 hours.
I added curl to dockerfile to install it.

## Additional notes or considerations:

I also added (with boy scout rule) a 3 min character length to the nlp processing, so we wouldn't process messages that are too short to be useful. (Entering "o" would give out biesdorf)