feat: api versions (#380)
This is critical as more and more services rely on this backend. And the
mobile apps new a review time until the changes will be rolled out and
we need to account for users with older versions of the mobile app.

This means if the api contract would change the app would be broken
until the users update it to the newest version. To prevent this we have
api versions.