chore: adapt api requests to explicit version (#381)
even though they would be routed to the latest version and thus not
breaking anything it is better to be explicit about it as to avoid
issues by forgetting to specify the version.

Every endpoint will have its own version to stay more modular that is
why I did not create a global variable.