refactor: `/risk-prediction/segment-colors` also returns risk value (#385)
## Describe the issue:
We are planning to implement a feature that will help users navigate to
their goals with the lowest possible risk. To do this, we need to
determine the risk level for each segment accurately. Currently, the
`/segment-colors/` endpoint only returns colors, which are aggregates of
the actual risk values. This would make the navigation feature
significantly less accurate.

## Explain how you solved the issue:
I modified the risk model to include the exact risk value along with the
calculated color for each segment. For example:

```json
{"segments_risk":{"S1-14":{"color":"#F05044","risk":0.861}...}
```
I updated version `v0` to serialize the data back to the old format for
backward compatibility. Additionally, I created a new version, `v1`,
which supports the segment risk values.

The web app has also been adapted to use the latest version of the API.

## Additional notes or considerations:
- I removed the `last_modified` field from the newest version of the
API. Since we were not using `If-Modified-Since` caching and it was only
included for backward compatibility, it was deemed unnecessary.
- We need to think about how we are rolling this out, to avoid version
conclicts with the mobile app as most users wont be on a version with
the explicit api versions.