More configurable (#412)
## Describe the issue:

The setup of freifahren for a new contributor was time-consuming, as
there have been intricate details not really documented or just not
stated. Some configurations for the Map.tsx were specific on Berlin. The
NLP code was outdated and had to be made better.

## Explain how you solved the issue:

Added dockerization for each of the services.
Added configuration for the map boundaries in the envs.
Beautified the nlp code, parting the aspects of the nlp service to
preparation functions and extraction functions.

## Additional notes or considerations: