feat: mobile app marketing popup (#423)
## Describe the issue:
The web app has a high number of users, while the mobile app has a low
number of users. We can convert the web app users to more longterm users
if they download the mobile app. To achieve this we should advertise the
mobile app inside of the web app.

## Explain how you solved the issue:
I added a pop up that will show up every 3 days and 30 seconds after the
user has been on the site.

![PHOTO-2025-05-18-14-41-34](https://github.com/user-attachments/assets/be487e39-9c49-4323-967a-1f91df52a2b0)


## Additional notes or considerations:
I refactored a bit to include a hook that now serves as a timing engine
for timed modals.