Telegram mini web app to report inspectors fre 473 (#419)
## Describe the issue:

The telegram group uses natural language to report, which is harder to
get information from. To provide more structured concise data and
comfortable usage,
we have added a Telegram Mini App to the Telegram Bot, that reflects the
FreiFahren WebApp

## Explain how you solved the issue:

Added a new package `mini_app` as a react structure to deploy a mini app
- near replica of the current report form found in the webapp

Added an author id to inform the backend the report was made through the
mini app.
Added an automatic /start command the telegram bot, to offer the service
when a new chat has been started


## Additional notes or considerations:
This mini app has to be connected with the telegram bot and environment
variables have to configured for production.