style: rename route
Since we are already breaking the existing contract by using `/basics/reports/` instead of `/basics/inspectors/`, we can further simplify the structure by removing the `/basics/` prefix and replacing it with versioning.

Adding versioning at this stage makes sense because it ensures that app requests are version-specific, eliminating the need to deprecate old routes once versioning is introduced.