feat: `GET transit/stations` (#437)
## Description:
I added the endpoint: `api.freifahren.org/transit/stations/` It returns
the entire stations data in this format:
```
type StationId = string
type LineId = string

type Stations = Record<StationId, Station>
type Station = {
    name: string
    coordinates: { latitude: number; longitude: number }
    lines: LineId[]
}
```
I also moved the endpoint to be a sub route of `/transit/` so that the
`getStations` and `getLines` can be handled in one service. Otherwise we
the service structure would drift from the API structure. Let me know
what you think of this.

## Implementation
The entire serialization step will be done at database level in a single
transaction. This ensures we have maximum performance. It might be a
little over the top since we will cache this any ways in the future
(Meaning the db calls runs only once). Let me know if I should revert to
the state from:
[65141c2](https://github.com/FreiFahren/FreiFahren/pull/437/commits/65141c2db88f623f5b6abdc5d8cecf0753cdbfe4)
since that implementation is slower but easier to read.

I would also like to get some feedback on how the types are created.
@mclrc is there a way how we can a better job at inferring this from
some db schema?

## Additional Notes
In order to not bloat this PR I will:
- add server side caching in a later PR
- add tests (not sure if we need tests for this endpoint, would like
some feedback)