feat: test `from` and `to` query param logic (#438)
## Describe the issue:
We implemented the `from` and `to` query params without testing it.

## Explain how you solved the issue:
I added tests that will test the following cases:
- `returns data in the specified timeframe when from and to query params
are present`
- `returns data in the standard timeframe when query params are missing`
- `returns 400 when from is after to`
- `returns 400 when timeframe is longer than the maximum'`

I also refactored the report service to make use of some constants, so
that these constants can also be used in the test. This will make sure
that when we change the default timeframe we won't have to update the
test with it.

All of this will now be run in CI, where we are testing against a proper
test db.

## Additional notes or considerations:
- I know we decided that we wanted to mock the DB, but we decided to do
this because we wanted a solution that is as simple as possible. While
implementing I realized that proper db mocks are much more complex than
creating a new test db. That is why I created the file
`01_create_dbs.sql`, which will be run by postgres automatically. This
creates the test db which is a copy of the freifahren db. Let me know
what you think?
- I had to alter the creation of the db to use the new format since we
do not want to use migrations until release
- I removed the go test CI since it kept breaking due to the recent
change and it is not needed anymore.