feat: notify telegram about new report + error handling (#440)
## Describe the issue:
After each new report created either from the web app or the mobile app,
we should send a request to the Telegram bot so that it can post a
message in the Telegram chat.
Additionally, we previously had no error handling in place.

## Explain how you solved the issue:
### Telegram Notification 
When an insert into the database is successful, we now send a `POST`
request to the Telegram bot to notify it about the new report. To
accomplish this, I enabled the source field to be sent via HTTP. This
approach naively trusts the API user to provide the correct source (I
added a test case to ensure it still defaults to telegram).
The alternative would be to parse the user agent on the backend to
determine the source without user input. However, this approach is
trickier to implement reliably, so I propose we stick to the naive
version for now until it becomes insufficient.

The way we currently construct the API request for Telegram is a bit
odd. I added a comment to remind us that this should be improved in the
long run.

### Error handling
Since this feature requires interaction with multiple services, it is
more prone to failure than simpler functionality. To mitigate this, we
need proper error handling. In this PR, I introduced a general structure
for handling HTTP errors, inspired by the way Go deals with errors (a
method that may not look pretty but encourages developers to treat
safety as a first-class concern).

The basic idea is that our service function returns an error of type
`ServerErrorBody`:
```
export interface ServerErrorDetails {
    internal_code: InternalCode
    description?: string
}

export interface ServerErrorBody {
    message: string
    details: ServerErrorDetails
}
```
The route function checks whether an error occurred. If so, it logs the
error and returns it. Otherwise, it responds with `status 200`:
```
        const error = await reportsService.createReport({
            ...reportData,
            source: reportData.source ?? 'telegram',
        })

        if (error !== null) {
            console.error(`Failed to create report: ${error.details.description}`)
            return c.json(error, 500)
        }

        return c.json(
            await reportsService.getReports({
                from: DateTime.now().minus({ hours: 1 }),
                to: DateTime.now(),
            })
        )
```

The purpose of this error type is to provide an `internal_code` that
uniquely identifies the error more specifically than standard HTTP
status codes like `500`. We can use this code on the client side to
match and display a user-friendly error message.
We also provide a `message` and a `description`: the message serves as a
heading, while the description gives additional context, mainly for
developers to debug issues more easily.

This level of detailed error handling is necessary due to recent user
feedback, as some users complained that it was not transparent why a
report failed.